+++ package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@fastify/forwarded": "2.1.0",
     "archiver": "5.3.1",
+    "fastify-plugin": "^3.0.1",
     "fluent-json-schema": "3.1.0",
     "geoip-lite": "1.4.2",
     "graasp-mailer": "github:graasp/graasp-mailer",
@@ -57,8 +58,5 @@
     "ts-node": "10.4.0",
     "typescript": "4.0.5"
   },
-  "packageManager": "yarn@3.2.1",
-  "resolutions": {
-    "@graasp/sdk": "portal:/Users/kim/git/graasp-utils"
-  }
+  "packageManager": "yarn@3.2.1"
 }
