+++ src/plugin.ts
@@ -1,22 +1,13 @@
 import { FastifyPluginAsync } from 'fastify';
+import fp from 'fastify-plugin';
 
-import {
-  Actor,
-  FileItemType,
-  Hostname,
-  LocalFileConfiguration,
-  S3FileConfiguration,
-} from '@graasp/sdk';
+import { Hostname } from '@graasp/sdk';
 
 import { ActionService } from './services/action/db-service';
 import { ActionTaskManager } from './services/action/task-manager';
 
 export interface GraaspActionsOptions {
-  graaspActor: Actor;
-  shouldSave?: boolean;
   hosts: Hostname[];
-  fileItemType: FileItemType;
-  fileConfigurations: { s3: S3FileConfiguration; local: LocalFileConfiguration };
 }
 
 const plugin: FastifyPluginAsync<GraaspActionsOptions> = async (fastify, options) => {
@@ -25,10 +16,13 @@ const plugin: FastifyPluginAsync<GraaspActionsOptions> = async (fastify, options
   const actionService = new ActionService();
   const actionTaskManager = new ActionTaskManager(actionService, hosts);
 
-  fastify.decorate('action', {
+  fastify.decorate('actions', {
     taskManager: actionTaskManager,
     dbService: actionService,
   });
 };
 
-export default plugin;
+export default fp(plugin, {
+  fastify: '3.x',
+  name: 'graasp-plugin-actions',
+});
