+++ test/app.ts
@@ -1,8 +1,16 @@
 import fastify from 'fastify';
 
-import { Item, ItemMembershipTaskManager, MemberTaskManager, TaskRunner } from '@graasp/sdk';
+import {
+  ActionService,
+  Item,
+  ItemMembershipTaskManager,
+  MemberTaskManager,
+  TaskRunner,
+} from '@graasp/sdk';
 import { ItemTaskManager } from 'graasp-test';
 
+import itemPlugin, { GraaspItemActionsOptions } from '../src/itemPlugin';
+import memberPlugin, { GraaspMemberActionsOptions } from '../src/memberPlugin';
 import plugin, { GraaspActionsOptions } from '../src/plugin';
 import { exportActionsInArchive } from '../src/utils/export';
 import { GRAASP_ACTOR } from './constants';
@@ -26,11 +34,12 @@ const schemas = {
 };
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-const build = async ({
+export const buildAppWithItemActions = async ({
   runner,
   itemTaskManager,
   itemMembershipTaskManager,
   memberTaskManager,
+  actionService,
   options,
   sendExportActionsEmail,
 }: {
@@ -38,7 +47,8 @@ const build = async ({
   itemTaskManager: ItemTaskManager;
   itemMembershipTaskManager: ItemMembershipTaskManager;
   memberTaskManager: MemberTaskManager;
-  options?: GraaspActionsOptions;
+  actionService: ActionService;
+  options?: GraaspItemActionsOptions;
   sendExportActionsEmail?: typeof exportActionsInArchive;
 }) => {
   const app = fastify();
@@ -59,8 +69,65 @@ const build = async ({
     sendExportActionsEmail: sendExportActionsEmail ?? jest.fn(),
   });
 
+  app.decorate('actions', {
+    dbService: actionService,
+  });
+
+  await app.register(itemPlugin, options);
+
+  return app;
+};
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const buildAppWithMemberActions = async ({
+  runner,
+  memberTaskManager,
+  actionService,
+  options,
+  sendExportActionsEmail,
+}: {
+  runner: TaskRunner<Item>;
+  memberTaskManager: MemberTaskManager;
+  actionService: ActionService;
+  options?: GraaspMemberActionsOptions;
+  sendExportActionsEmail?: typeof exportActionsInArchive;
+}) => {
+  const app = fastify();
+  app.addSchema(schemas);
+  app.decorateRequest('member', GRAASP_ACTOR);
+
+  app.decorate('taskRunner', runner);
+
+  app.decorate('members', {
+    taskManager: memberTaskManager,
+  });
+  app.decorate('mailer', {
+    sendExportActionsEmail: sendExportActionsEmail ?? jest.fn(),
+  });
+
+  app.decorate('actions', {
+    dbService: actionService,
+  });
+
+  await app.register(memberPlugin, options);
+
+  return app;
+};
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const buildApp = async ({
+  runner,
+  options,
+}: {
+  runner: TaskRunner<Item>;
+  options?: GraaspActionsOptions;
+}) => {
+  const app = fastify();
+  app.addSchema(schemas);
+
+  app.decorate('taskRunner', runner);
+
   await app.register(plugin, options);
 
   return app;
 };
-export default build;
